+++ selfdrive/controls/lib/lane_planner.py
@@ -98,7 +98,8 @@ class LaneOffset:
   AUTO_ENABLE_TRAFFIC_MIN_SPEED = 10. * CV.MPH_TO_MS
   AUTO_ENABLE_MIN_SPEED_DEADZONE = 2. # [m/s]
   
-  AUTO_TRAFFIC_TIMEOUT = 30. # [s] amount of time auto lane position will be kept after last car is seen
+  AUTO_TRAFFIC_TIMEOUT_ONCOMING = 60. # [s] amount of time auto lane position will be kept after last car is seen
+  AUTO_TRAFFIC_TIMEOUT_ONGOING = 5. # [s] amount of time auto lane position will be kept after last car is seen
   AUTO_TRAFFIC_MIN_TIME = 1.1 # [s] time an oncoming/ongoing car needs to be observed before it will be taken to indicate traffic
   AUTO_TRAFFIC_MIN_SPEED = 9. # [m/s] need to go faster than this to be considered moving
   AUTO_CUTOFF_STEER_ANGLE = 110. # [degrees] auto lane position reset traffic monitoring after this
@@ -134,6 +135,9 @@ class LaneOffset:
     self._auto_auto_enabled = False
     self._v_ego_last = 0.
     
+    self.offset_min = -self.OFFSET_MAX
+    self.offset_max = self.OFFSET_MAX
+    
     self._left_traffic = LANE_TRAFFIC.NONE
     self._right_traffic = LANE_TRAFFIC.NONE
     self._left_traffic_last = LANE_TRAFFIC.NONE
@@ -146,6 +150,8 @@ class LaneOffset:
     self._right_traffic_min_sep_dist = FirstOrderFilter(self.ADJACENT_TRAFFIC_SEP_DIST_NONE, 10.0, DT_MDL)
     self._left_traffic_last_seen_t = 0.
     self._right_traffic_last_seen_t = 0.
+    self._left_traffic_last_seen = LANE_TRAFFIC.NONE
+    self._right_traffic_last_seen = LANE_TRAFFIC.NONE
     self._left_traffic_temp_t = 0.
     self._right_traffic_temp_t = 0.
     self._lprob_last = 0.
@@ -176,29 +182,12 @@ class LaneOffset:
             ret = AUTO_AUTO_LANE_MODE.DISENGAGE
             self._auto_auto_enabled = False
     return ret
-
-  def do_auto_enable_lanelines(self, ret, road_type, v_ego):
-    v_ego_diff = apply_deadzone(v_ego - self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED, self.AUTO_ENABLE_MIN_SPEED_DEADZONE)
-    v_ego_diff_last = apply_deadzone(self._v_ego_last - self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED, self.AUTO_ENABLE_MIN_SPEED_DEADZONE)
-    if road_type != self._road_type_last or ret == AUTO_AUTO_LANE_MODE.DISENGAGE \
-        or (v_ego_diff > 0. and v_ego_diff_last <= 0.) \
-        or (v_ego_diff < 0. and v_ego_diff_last >= 0.):
-      if (road_type in self.AUTO_ENABLE_ROAD_TYPES and v_ego_diff > 0. and (not self._auto_is_active or ret == AUTO_AUTO_LANE_MODE.DISENGAGE)):
-        ret = AUTO_AUTO_LANE_MODE.NO_CHANGE if ret == AUTO_AUTO_LANE_MODE.DISENGAGE else AUTO_AUTO_LANE_MODE.ENGAGE
-        self._auto_auto_enabled = True
-      elif self._auto_is_active and self._auto_auto_enabled \
-        and (road_type not in self.AUTO_ENABLE_ROAD_TYPES or v_ego_diff < 0.) :
-          ret = AUTO_AUTO_LANE_MODE.DISENGAGE
-          self._auto_auto_enabled = False
-    return ret
             
   def do_auto_enable(self, road_type):
     ret = AUTO_AUTO_LANE_MODE.NO_CHANGE
     v_ego = self._cs.vEgo if self._cs is not None else 0.
     
     ret = self.do_auto_enable_traffic(ret, v_ego)
-    if ret != AUTO_AUTO_LANE_MODE.ENGAGE:
-      ret = self.do_auto_enable_lanelines(ret, road_type, v_ego)
 
     self._auto_auto_lane_position_action = ret
     self._lprob_last = self._lane_probs[1]
@@ -250,10 +239,11 @@ class LaneOffset:
     return
   
   def update_traffic_info(self, rs, lane_width, md):
-    left_traffic = LANE_TRAFFIC.NONE
-    right_traffic = LANE_TRAFFIC.NONE
     if self._cs is None or abs(self._cs.steeringAngleDeg) > self.TRAFFIC_NEW_DETECT_STEER_CUTOFF:
       return
+    
+    left_traffic = LANE_TRAFFIC.NONE
+    right_traffic = LANE_TRAFFIC.NONE
     
     sep_dist = self.ADJACENT_TRAFFIC_SEP_DIST_NONE # [s] if <= 1 adjacent cars, this is the "separation" between them
     
@@ -273,7 +263,9 @@ class LaneOffset:
         lv = [l.vLeadK for l in l1]
         min_v, max_v, mean_v = min_max_mean(lv)
         check_v = min_v if abs(min_v - mean_v) < abs(max_v - mean_v) else max_v
-        if check_v > self.AUTO_TRAFFIC_MIN_SPEED and self._lane_width_mean_left_adjacent > 0.:
+        if check_v > self.AUTO_TRAFFIC_MIN_SPEED \
+            and self._lane_probs[0] > self.AUTO_MIN_ADJACENT_LANELINE_PROB \
+            and self._lane_width_mean_left_adjacent > 0.:
           left_traffic = LANE_TRAFFIC.ONGOING
           if len(l1) > 1:
             l1.sort(key= lambda x:x.dRel)
@@ -304,7 +296,9 @@ class LaneOffset:
         lv = [l.vLeadK for l in l1]
         min_v, max_v, mean_v = min_max_mean(lv)
         check_v = min_v if abs(min_v - mean_v) < abs(max_v - mean_v) else max_v
-        if check_v > self.AUTO_TRAFFIC_MIN_SPEED and self._lane_width_mean_right_adjacent > 0.:
+        if check_v > self.AUTO_TRAFFIC_MIN_SPEED \
+            and self._lane_probs[3] > self.AUTO_MIN_ADJACENT_LANELINE_PROB \
+            and self._lane_width_mean_right_adjacent > 0.:
           right_traffic = LANE_TRAFFIC.ONGOING
           if len(l1) > 1:
             l1.sort(key= lambda x:x.dRel)
@@ -323,9 +317,14 @@ class LaneOffset:
         self._left_traffic_temp_t = self._t
       elif self._t - self._left_traffic_temp_t >= self.AUTO_TRAFFIC_MIN_TIME:
         self._left_traffic_last_seen_t = self._t
+        self._left_traffic_last_seen = left_traffic
         self._left_traffic = left_traffic
-    elif self._t - self._left_traffic_last_seen_t > self.AUTO_TRAFFIC_TIMEOUT:
-      self._left_traffic = left_traffic
+    else:
+      if (self._left_traffic_last_seen == LANE_TRAFFIC.ONCOMING \
+          and self._t - self._left_traffic_last_seen_t > self.AUTO_TRAFFIC_TIMEOUT_ONCOMING) \
+        or (self._left_traffic_last_seen != LANE_TRAFFIC.ONCOMING \
+          and self._t - self._left_traffic_last_seen_t > self.AUTO_TRAFFIC_TIMEOUT_ONGOING):
+        self._left_traffic = left_traffic
     
     self._left_traffic_temp = left_traffic
       
@@ -334,9 +333,14 @@ class LaneOffset:
         self._right_traffic_temp_t = self._t
       elif self._t - self._right_traffic_temp_t >= self.AUTO_TRAFFIC_MIN_TIME:
         self._right_traffic_last_seen_t = self._t
+        self._left_traffic_last_seen = right_traffic
         self._right_traffic = right_traffic
-    elif self._t - self._right_traffic_last_seen_t > self.AUTO_TRAFFIC_TIMEOUT:
-      self._right_traffic = right_traffic
+    else:
+      if (self._right_traffic_last_seen == LANE_TRAFFIC.ONCOMING \
+          and self._t - self._right_traffic_last_seen_t > self.AUTO_TRAFFIC_TIMEOUT_ONCOMING) \
+        or (self._right_traffic_last_seen != LANE_TRAFFIC.ONCOMING \
+          and self._t - self._right_traffic_last_seen_t > self.AUTO_TRAFFIC_TIMEOUT_ONGOING):
+        self._right_traffic = right_traffic
     
     self._right_traffic_temp = right_traffic
       
@@ -355,15 +359,6 @@ class LaneOffset:
           and self._left_traffic == LANE_TRAFFIC.NONE:
         lane_pos_auto = 1.
         timeout_override = True
-      elif self._lane_probs[1] > self.AUTO_MIN_LANELINE_PROB \
-          and self._lane_probs[2] > self.AUTO_MIN_LANELINE_PROB:
-        if (self._lane_width_mean_left_adjacent > 0. and self._lane_width_mean_right_adjacent > 0.) \
-            or (self._lane_width_mean_left_adjacent == 0. and self._lane_width_mean_right_adjacent == 0.):
-          lane_pos_auto = 0.
-        elif self._lane_width_mean_left_adjacent > 0. and self._shoulder_width_mean_right >= self.AUTO_MIN_SHOULDER_WIDTH_FACTOR * lane_width:
-          lane_pos_auto = -1.
-        elif self._lane_width_mean_right_adjacent > 0. and self._shoulder_width_mean_left >= self.AUTO_MIN_SHOULDER_WIDTH_FACTOR * lane_width:
-          lane_pos_auto = 1.
     if lane_pos_auto != 0. \
         and ((self._lat_accel_cur >= self.AUTO_MAX_CUR_LAT_ACCEL \
           and np.sign(self._lat_curvature_cur) == np.sign(lane_pos_auto)) \
@@ -407,26 +402,33 @@ class LaneOffset:
           lane_depart = l_lane_depart or r_lane_depart
           
         if abs(self._cs.steeringAngleDeg) > self.AUTO_CUTOFF_STEER_ANGLE or lane_depart:
-          self._left_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT + 1
-          self._right_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT + 1
+          self._left_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT_ONCOMING + 1
+          self._right_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT_ONCOMING + 1
           self._right_traffic = LANE_TRAFFIC.NONE
           self._left_traffic = LANE_TRAFFIC.NONE
+        else:
+          if self._left_traffic == LANE_TRAFFIC.ONCOMING and \
+            (md.laneLines[1].y[0] > -(0.5 + self.offset) or md.laneLines[1].y[0] < -(2.5 + self.offset)):
+            self._left_traffic = LANE_TRAFFIC.NONE
+          if self._right_traffic == LANE_TRAFFIC.ONCOMING and \
+            (md.laneLines[2].y[0] < (0.5 + self.offset) or md.laneLines[2].y[0] > (2.5 + self.offset)):
+            self._right_traffic = LANE_TRAFFIC.NONE 
           
         if self._right_traffic == LANE_TRAFFIC.ONGOING and self._lane_width_mean_right_adjacent == 0.:
           self._right_traffic = LANE_TRAFFIC.NONE
-          self._right_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT + 1
+          self._right_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT_ONCOMING + 1
           
         if self._left_traffic == LANE_TRAFFIC.ONGOING and self._lane_width_mean_left_adjacent == 0.:
           self._left_traffic = LANE_TRAFFIC.NONE
-          self._left_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT + 1
+          self._left_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT_ONCOMING + 1
           
         if self._lat_plan.lProb < 0.2 and self._left_traffic == LANE_TRAFFIC.NONE \
             and self._t - self._lane_state_changed_last_t < self.AUTO_LANE_STATE_MIN_TIME:
-          self._left_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT + 1
+          self._left_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT_ONCOMING + 1
           
         if self._lat_plan.rProb < 0.2 and self._right_traffic == LANE_TRAFFIC.NONE \
             and self._t - self._lane_state_changed_last_t < self.AUTO_LANE_STATE_MIN_TIME:
-          self._right_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT + 1
+          self._right_traffic_last_seen_t -= self.AUTO_TRAFFIC_TIMEOUT_ONCOMING + 1
           
         self.update_lane_pos_auto(lane_width)
       if self._long_plan is not None:
@@ -450,6 +452,18 @@ class LaneOffset:
     else:
       self.lane_pos = lane_pos
     offset = self.OFFSET * self.offset_scale * self.lane_pos * lane_width * (self.AUTO_OFFSET_FACTOR if self._auto_is_active else 1.)
+    
+    if self._left_traffic_temp == LANE_TRAFFIC.ONCOMING:
+      self.offset_max = 0.0
+    else:
+      self.offset_max = self.OFFSET_MAX
+    if self._right_traffic_temp == LANE_TRAFFIC.ONCOMING:
+      self.offset_min = 0.0
+    else:
+      self.offset_min = -self.OFFSET_MAX
+    
+    offset = clip(offset, self.offset_min, self.offset_max)
+    
     if offset > self.offset:
       self.offset = min(offset, self.offset + (self.STEP_SLOW if do_slow else self.STEP))
     elif offset < self.offset:
@@ -466,6 +480,7 @@ class LanePlanner:
     self.lane_width_certainty = FirstOrderFilter(1.0, 0.95, DT_MDL)
     self.lane_width = LANE_WIDTH_DEFAULT
     self.lane_offset = LaneOffset(mass)
+    self.lane_dist_from_center = FirstOrderFilter(0.0, 5.0, DT_MDL)
 
     self.lll_prob = 0.
     self.rll_prob = 0.
@@ -487,8 +502,11 @@ class LanePlanner:
       # left and right ll x is the same
       self.ll_x = md.laneLines[1].x
       # only offset left and right lane lines; offsetting path does not make sense
-      self.lll_y = np.array(md.laneLines[1].y) - self.camera_offset - offset
-      self.rll_y = np.array(md.laneLines[2].y) - self.camera_offset - offset
+      self.lll_y = np.array(md.laneLines[1].y) - self.camera_offset
+      self.rll_y = np.array(md.laneLines[2].y) - self.camera_offset
+      self.lane_dist_from_center.update((self.rll_y[0] + self.lll_y[0]) * 0.5)
+      self.lll_y -= offset
+      self.rll_y -= offset
       self.lll_prob = md.laneLineProbs[1]
       self.rll_prob = md.laneLineProbs[2]
       self.lll_std = md.laneLineStds[1]