+++ selfdrive/controls/lib/longitudinal_planner.py
@@ -65,8 +65,8 @@ def calc_cruise_accel_limits(v_ego, following, accelMode):
     a_cruise_max = interp(v_ego, _A_CRUISE_MAX_BP, _A_CRUISE_MAX_V_MODE_LIST[accelMode])
   return [a_cruise_min, a_cruise_max]
 
-LEAD_ONE_PLUS_TR_BUFFER = 0.1 # [s] follow distance between lead and lead+1 to run the lead+1 mpc
-LEAD_ONE_PLUS_STOPPING_DISTANCE_BUFFER = 2.0 # [m]
+LEAD_ONE_PLUS_TR_BUFFER = 0.0 # [s] follow distance between lead and lead+1 to run the lead+1 mpc
+LEAD_ONE_PLUS_STOPPING_DISTANCE_BUFFER = 1.0 # [m]
 
 def limit_accel_in_turns(v_ego, angle_steers, a_target, CP):
   """
@@ -159,6 +159,8 @@ class Planner():
     self.lead_0 = sm['radarState'].leadOne
     self.lead_1 = sm['radarState'].leadTwo
     self.lead_0_plus = sm['radarState'].leadOnePlus
+    
+    self.mpcs['lead0'].df.lateralPlan = sm['lateralPlan']
 
     enabled = (long_control_state == LongCtrlState.pid) or (long_control_state == LongCtrlState.stopping)
     following = self.lead_0.status and self.lead_0.dRel < 45.0 and self.lead_0.vLeadK > v_ego and self.lead_0.aLeadK > 0.0
@@ -220,11 +222,7 @@ class Planner():
     self.lead_accel = np.inf
     for key in self.mpcs:
       if key == 'lead0p1':
-        if self.lead_0_plus.status and self.lead_0.status:
-          tr = self.mpcs['lead0'].tr + LEAD_ONE_PLUS_TR_BUFFER
-          self.mpcs['lead0p1'].tr_override = True
-          self.mpcs['lead0p1'].tr = tr
-        else:
+        if not self.lead_0_plus.status or not self.lead_0.status:
           continue
       self.mpcs[key].set_cur_state(self.v_desired, self.a_desired)
       self.mpcs[key].update(sm['carState'], sm['radarState'], v_cruise, a_mpc[key], active_mpc[key])