+++ selfdrive/controls/lib/turn_speed_controller.py
@@ -70,7 +70,7 @@ class TurnSpeedController():
 
     self._next_speed_limit_prev = 0.
 
-    self._a_target = FirstOrderFilter(0., 1.0, DT_MDL)
+    self._a_target = FirstOrderFilter(0., 2.0, DT_MDL)
 
   @property
   def a_target(self):