+++ selfdrive/controls/lib/vision_turn_controller.py
@@ -32,7 +32,7 @@ _A_LAT_REG_MAX = 1.55  # Maximum lateral acceleration
 
 # Lookup table for the minimum smooth deceleration during the ENTERING state
 # depending on the actual maximum absolute lateral acceleration predicted on the turn ahead.
-_ENTERING_SMOOTH_DECEL_V = [0.0, -0.2, -2.0]  # min decel value allowed on ENTERING state
+_ENTERING_SMOOTH_DECEL_V = [0.0, -0.2, -1.5]  # min decel value allowed on ENTERING state
 _ENTERING_SMOOTH_DECEL_BP = [1.2, 1.35, 2.3]  # absolute value of lat acc ahead
 
 # Lookup table for the acceleration for the TURNING state
@@ -113,7 +113,7 @@ class VisionTurnController():
     self._v_cruise_setpoint = 0.
     self._v_ego = 0.
     self._a_ego = 0.
-    self._a_target = FirstOrderFilter(0., 1.0, DT_MDL)
+    self._a_target = FirstOrderFilter(0., 2.0, DT_MDL)
     self._v_overshoot = 0.
     self._state = VisionTurnControllerState.disabled
     self._CS = None